Refactor .gitignore to speed up 'git status'
Various threads on the git mailing list suggest that 'git status' (and
other commands which stat the entire working tree) run faster when a
top-level .gitignore file is split into many smaller pieces.  In
particular, wildcard rules in a top-level .gitignore file are supposedly
undesirable.

This seems to be a little faster.  I haven't tested this on a Windows
machine, so the new rules may fail to ignore a few files.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29253 b24aca02-f418-0410-9c49-835e4aa915c3