Add new console commands for teleporting and getting position
The new commands are:
  pos: Print x, y, z and yaw
  teleport: Zap player to specified x, y, z and yaw

This teleport command is for debugging purposes only--it includes visual
distortions and sound effects, and we don't understand all the details of
how it interacts with the underlying system.

We also fixed some incorrect documentation on yaw and pitch in the watchdir
command.

Paired-with: Brian Campbell <brian.p.campbell@dartmouth.edu>
Bug: http://iml.dartmouth.edu/fogbugz/default.php?11221